refactor: introduce TypeMinimization.scala (#4314)
