+++ main/src/ca/uwaterloo/flix/language/phase/extra/CodeHinter.scala
@@ -21,7 +21,7 @@ import ca.uwaterloo.flix.language.ast.TypedAst.Predicate.{Body, Head}
 import ca.uwaterloo.flix.language.ast.TypedAst._
 import ca.uwaterloo.flix.language.ast.{Ast, SourceLocation, Symbol, Type, TypeConstructor, TypedAst}
 import ca.uwaterloo.flix.language.errors.CodeHint
-import ca.uwaterloo.flix.language.phase.unification.BoolTable
+import ca.uwaterloo.flix.language.phase.unification.TypeMinimization
 
 object CodeHinter {
 
@@ -394,7 +394,7 @@ object CodeHinter {
       Nil
     } else {
       // Case 2: Formula is big. Try to minimize it.
-      val minType = BoolTable.minimizeType(tpe)
+      val minType = TypeMinimization.minimizeType(tpe)
       if (numberOfVarOccurs(minType) < 5) {
         // Case 2.1: Formula is small. Good.
         Nil
