+++ main/src/ca/uwaterloo/flix/language/phase/unification/BoolTable.scala
@@ -15,11 +15,9 @@
  */
 package ca.uwaterloo.flix.language.phase.unification
 
-import ca.uwaterloo.flix.api.Flix
-import ca.uwaterloo.flix.language.ast.{Kind, Symbol, Type, TypeConstructor}
 import ca.uwaterloo.flix.language.phase.unification.BoolFormula._
-import ca.uwaterloo.flix.util.{InternalCompilerException, LocalResource, StreamOps}
 import ca.uwaterloo.flix.util.collection.Bimap
+import ca.uwaterloo.flix.util.{InternalCompilerException, LocalResource, StreamOps}
 
 import java.io.IOException
 import java.util.zip.ZipInputStream
@@ -56,12 +54,12 @@ object BoolTable {
   /**
     * The size a formula (but represented as a type) must have before we try to minimize it.
     */
-  private val Threshold: Int = 10
+  val Threshold: Int = 10
 
   /**
     * A Boolean variable is represented by a unique number.
     */
-  private type Variable = Int
+  type Variable = Int
 
   /**
     * A table that maps Boolean semantic functions to their minimal formulas.
@@ -69,57 +67,6 @@ object BoolTable {
     * The table is pre-computed and initialized when this class is loaded.
     */
   private lazy val Table: Array[BoolFormula] = initTable()
-
-  /**
-    * Attempts to minimize the given Boolean formula `tpe`.
-    *
-    * Returns the same formula or a smaller formula that is equivalent.
-    */
-  def minimizeType(tpe0: Type)(implicit flix: Flix): Type = {
-    // Check whether minimization via tabling is disabled.
-    if (flix.options.xnobooltable) {
-      return tpe0
-    }
-
-    // Check that the `tpe` argument is a Boolean formula.
-    if (tpe0.kind != Kind.Bool && tpe0.kind != Kind.Effect) {
-      throw InternalCompilerException(s"Unexpected non-Bool/non-Effect kind: '${tpe0.kind}'.")
-    }
-
-    // Erase aliases to get a processable type
-    val tpe = Type.eraseAliases(tpe0)
-
-    // Compute the size of  `tpe`.
-    val currentSize = tpe.size
-
-    // Return `tpe` immediately if it is "small".
-    if (currentSize < Threshold) {
-      return tpe
-    }
-
-    // Compute the variables in `tpe`.
-    val tvars = tpe.typeVars.toList.map(tvar => BoolFormula.VarOrEff.Var(tvar.sym))
-    val effs = getEffects(tpe).toList.map(BoolFormula.VarOrEff.Eff)
-
-    // Construct a bi-directional map from type variables to indices.
-    // The idea is that the first variable becomes x0, the next x1, and so forth.
-    val m = (tvars ++ effs).zipWithIndex.foldLeft(Bimap.empty[BoolFormula.VarOrEff, Variable]) {
-      case (macc, (sym, x)) => macc + (sym -> x)
-    }
-
-    // Convert the type `tpe` to a Boolean formula.
-    val input = tpe.kind match {
-      case Kind.Bool => fromBoolType(tpe, m)
-      case Kind.Effect => fromEffType(tpe, m)
-      case _ => throw InternalCompilerException(s"Unexpected non-Bool/non-Effect kind: '${tpe.kind}'.")
-    }
-
-    // Minimize the Boolean formula.
-    val minimized = minimizeFormula(input)
-
-    // Convert the formula back to a type.
-    toType(minimized, m, tpe.kind, tpe.loc)
-  }
 
   /**
     * Attempts to minimize the given formula `f`.
@@ -388,22 +335,4 @@ object BoolTable {
     */
   private def leftPad(s: String, len: Int): String =
     ' '.toString * (len - s.length()) + s
-
-  /**
-    * Gets all the effects in the given type.
-    */
-  private def getEffects(t: Type): SortedSet[Symbol.EffectSym] = t match {
-    case Type.Cst(TypeConstructor.Effect(sym), _) => SortedSet(sym)
-
-    case _: Type.Cst => SortedSet.empty
-    case _: Type.KindedVar => SortedSet.empty
-
-    case Type.Apply(tpe1, tpe2, loc) => getEffects(tpe1) ++ getEffects(tpe2)
-    case Type.Alias(cst, args, tpe, loc) => getEffects(tpe)
-
-    case _: Type.Ascribe => throw InternalCompilerException("Unexpected unkinded type.")
-    case _: Type.UnkindedVar => throw InternalCompilerException("Unexpected unkinded type.")
-    case _: Type.UnkindedArrow => throw InternalCompilerException("Unexpected unkinded type.")
-    case _: Type.ReadWrite => throw InternalCompilerException("Unexpected unkinded type.")
-  }
 }
